refactor: split example/all_types/all_types.go — WKB helpers into wkb.go
